Spring JPA. Fix commit

+++ service/src/main/java/com/viktoria/spring/database/repository/ClaimRepository.java
@@ -3,6 +3,6 @@ package com.viktoria.spring.database.repository;
 import com.viktoria.spring.database.entity.Claim;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClaimRepository extends JpaRepository<Claim, Integer> {
+public interface ClaimRepository extends JpaRepository<Claim, Long> {
 
 }
