+++ service/src/main/java/com/viktoria/spring/database/repository/ExtrasRepository.java
@@ -3,6 +3,6 @@ package com.viktoria.spring.database.repository;
 import com.viktoria.spring.database.entity.Extras;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExtrasRepository extends JpaRepository<Extras, Integer> {
+public interface ExtrasRepository extends JpaRepository<Extras, Long> {
 
 }
